requirements.txt
azure-functions
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-mgmt-videoanalyzer>=1.0.0
requests>=2.31.0
openai>=1.12.0
python-docx>=1.1.0
Pillow>=10.2.0
azure-communication-email>=1.0.0
tenacity>=8.2.0
